Factored terminal details into own headers.
